+++ example/lib/pickFromURL/pick_from_url_view.dart
@@ -28,7 +28,6 @@ class PickFromURLView extends StatelessWidget {
         body: SafeArea(
           child: Center(
             child: FlutterAVPlayerView(
-              // filePath: 'assets/videos/butterfly.mp4',
               urlString:
                   'https://flutter.github.io/assets-for-api-docs/assets/videos/butterfly.mp4',
             ),